Add files via upload
Chess.py runs a very basic chess game in command line or your interpreter. No fancy graphics unfortunately.
The code is mostly in English but parts are in Dutch as this was an assignment for school and we had to use certain function names. I'm in the process of translating the entire code into English.